Anti Gravity/requirements.txt
fastapi
uvicorn
python-telegram-bot
SQLAlchemy
python-dotenv
aiohttp
replicate
jinja2
pydantic
requests
